pytest: Separate policy tests
